+++ R/mgidi.R
@@ -16,10 +16,10 @@
 #' genotype with the lowest MGIDI is then closer to the ideotype and therefore
 #' presents desired values for all the analyzed traits.
 #'
-#' @param .data An object fitted with the function \code{\link{gamem}} or a
-#'   two-way table with BLUPs for genotypes in each trait (genotypes in rows and
-#'   traits in columns). In the last case, row names must contain the genotypes
-#'   names.
+#' @param .data An object fitted with the function \code{\link{gafem}()},
+#'   \code{\link{gamem}()} or a two-way table with BLUPs for genotypes in each
+#'   trait (genotypes in rows and traits in columns). In the last case, row
+#'   names must contain the genotypes names.
 #' @param SI An integer (0-100). The selection intensity in percentage of the
 #' total number of genotypes.
 #' @param mineval The minimum value so that an eigenvector is retained in the
@@ -30,8 +30,9 @@
 #'   variables in which lower values are desired. For example, \code{ideotype =
 #'   c("h, h, h, h, l")} will consider that the ideotype has higher values for
 #'   the first four traits and lower values for the last trait. If \code{.data}
-#'   is a model fitted with the function \code{\link{gamem}}, the order of the
-#'   traits will be the declared in the argument \code{resp} in that function.
+#'   is a model fitted with the functions \code{\link{gafem}()} or
+#'   \code{\link{gamem}()}, the order of the traits will be the declared in the
+#'   argument \code{resp} in those functions.
 #' @param use The method for computing covariances in the presence of missing
 #'   values. Defaults to \code{complete.obs}, i.e., missing values are handled
 #'   by casewise deletion.